ENH: Convert to autopep8 compliant
Used mechanism to make the code consistent with pep8 standards
to be compliant with nipype and other standards.

This was prompted by warnings given for inconsistent indentation
warnings by the editor.  There were many places where tab indentation
was followed by space indentation.

This change also makes the code easier to read for those of us
are used to nipype and strict pep8 compliant.