ENH: Require `NumPy` greater or equal to 1.26.0
Require `NumPy` greater or equal to 1.26.0: 1.26.0 is the first release
that supported Python 3.12 and supports versions 3.9-3.12.

No longer restrict `NumPy` versions to the 1.x series.

Documentation:
https://github.com/numpy/numpy/releases/tag/v1.26.0

Require `NiBabel` greater or equal to 4.0.0: the old `trackvis` API
existing in previous versions was removed starting from version 4.0.0.
`tract_querier` reliance on the old `trackvis` API was removed in commit
813b570.

The 4.0.0 version release notes can be read in:
https://github.com/nipy/nibabel/releases/tag/4.0.0

No longer restrict VTK versions to the 8.x series.

Add Python 3.12 and 3.13 to the GHA package testing matrix. For Python
3.13 exclude the `minimal` dependency configuration as some are not
compatible with it.