ENH: Do not import `range` from `six`
Do not import `range` from `six`: `six` is a Python 2 and 3
compatibility library. Since support for Python 2 was dropped following
the discussion in https://github.com/demianw/tract_querier/pull/53, this
patch set removes `six` as a dependency of the project.

The built-in Python `range()` is equivalent to the `six.moves.range`
method, so adopt the former by avoiding to import it from `six.moves`.

Similarly, use `range()` instead of `xrange()`, which is only supported
in Python 2. Both are equivalent.

Documentation:
https://pypi.org/project/six/