DOC: Update package versions in installation instructions (#92)
Update package version in installation instructions.

After commits 813b570 and e72d4e3 were merged, the minimum `NiBabel`
version was raised to 4.0.0 and the restriction on newer versions was
removed, and the minimum `NumPy` version was raised to 1.26 and newer
versions were allowed.

This enabled to use Python version greater than 3.12.